moved latticemaze and plot_dataset
