+++ maze_dataset/generation/generators.py
@@ -4,7 +4,7 @@ from typing import Any, Callable
 import numpy as np
 
 from maze_dataset.constants import CoordArray
-from maze_dataset.generation.lattice_maze import (
+from maze_dataset.maze.lattice_maze import (
     NEIGHBORS_MASK,
     ConnectionList,
     Coord,
