+++ maze_dataset/maze/lattice_maze.py
