+++ maze_dataset/plotting/plot_dataset.py
