+++ maze_dataset/plotting/plot_maze.py
@@ -12,7 +12,7 @@ from matplotlib.colors import ListedColormap, Normalize
 from muutils.tensor_utils import NDArray
 
 from maze_dataset.constants import Coord, CoordArray, CoordList
-from maze_dataset.generation.lattice_maze import (
+from maze_dataset.maze.lattice_maze import (
     Coord,
     CoordArray,
     LatticeMaze,
